traefik/RateLimit/server/requirements.txt
python-dotenv==1.0.1
fastapi==0.111.1 